\name{Indices_0809}
     \docType{data}
     \alias{Indices_0809}
     \title{Relative losses in percent on 3 stock market indices}
     \description{
       Relative losses in percent on 3 stock market indices in the years 2008 and 2009.
     }
     \usage{Indices_0809}
     \format{A vector containing 477 observations.}
     \source{http://www.uni-koeln.de/wiso-fak/wisostatsem/Forschung/WMT_Regions.htm}
       \keyword{datasets}